Consolidating all MongoDB related things in one place - all up and running conceptually

+++ Source/Kernel/Events.Projections/Pipelines/ProjectionPipelineHandlerLogMessages.cs
@@ -8,15 +8,15 @@ namespace Cratis.Events.Projections.Pipelines
     public static partial class ProjectionPipelineHandlerLogMessages
     {
         [LoggerMessage(0, LogLevel.Debug, "Handling event with sequence number {SequenceNumber}")]
-        internal static partial void HandlingEvent(this ILogger logger, uint sequenceNumber);
+        internal static partial void HandlingEvent(this ILogger logger, ulong sequenceNumber);
 
         [LoggerMessage(1, LogLevel.Trace, "Getting initial values for event with sequence number {SequenceNumber}")]
-        internal static partial void GettingInitialValues(this ILogger logger, uint sequenceNumber);
+        internal static partial void GettingInitialValues(this ILogger logger, ulong sequenceNumber);
 
         [LoggerMessage(2, LogLevel.Trace, "Projecting for event with sequence number {SequenceNumber}")]
-        internal static partial void Projecting(this ILogger logger, uint sequenceNumber);
+        internal static partial void Projecting(this ILogger logger, ulong sequenceNumber);
 
         [LoggerMessage(3, LogLevel.Trace, "Saving result for event with sequence number {SequenceNumber}")]
-        internal static partial void SavingResult(this ILogger logger, uint sequenceNumber);
+        internal static partial void SavingResult(this ILogger logger, ulong sequenceNumber);
     }
 }
