+++ Source/Kernel/Events.Store.MongoDB/EventLogsLogMessages.cs
@@ -11,7 +11,7 @@ namespace Cratis.Events.Store.MongoDB
     public static partial class EventLogsLogMessages
     {
         [LoggerMessage(0, LogLevel.Information, "Appending event with '{SequenceNumber}' as sequence number")]
-        internal static partial void Appending(this ILogger logger, uint sequenceNumber);
+        internal static partial void Appending(this ILogger logger, ulong sequenceNumber);
 
         [LoggerMessage(1, LogLevel.Error, "Problem appending event to storage")]
         internal static partial void AppendFailure(this ILogger logger, Exception exception);
